Enable transparency setting
