+++ manifest.json
@@ -10,15 +10,14 @@
   },
   "content_scripts": [
     {
-      "matches": [
-        "*://www.nicovideo.jp/watch/*"
-      ],
-      "js": [
-        "scripts/index.js"
-      ]
+      "matches": ["*://www.nicovideo.jp/watch/*"],
+      "js": ["scripts/index.js"]
     }
   ],
-  "web_accessible_resources": ["scripts/hack_fetch_thread.js"],
+  "web_accessible_resources": [
+    "scripts/hack_fetch_thread.js",
+    "scripts/hack_comment_alpha.js"
+  ],
   "background": {
     "scripts": ["scripts/background.js"]
   },
