+++ webpack.config.js
@@ -3,6 +3,7 @@ module.exports = {
   entry: {
     'index': './index.js',
     'hack_fetch_thread': './hack_fetch_thread.js',
+    'hack_comment_alpha': './hack_comment_alpha.js',
     'background': './background.js',
     'popup': './popup.js'
   },
